First commi

+++ database/migrations/2023_10_10_113437_create_product_materials_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->foreignId("product_id")->constrained("products")->onDelete("cascade");
             $table->foreignId("material_id")->constrained("materials")->onDelete("cascade");
-            $table->unsignedBigInteger("quantity");
+            $table->decimal("quantity",5,1);
             $table->timestamps();
         });
     }
